+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Export `checkIterConstraints()`
 - Provide `AsIter` type
+- Impl `Iter.collectAlloc()`, `Iter.collectBuf()`, `Iter.collectArrayList()` and `Iter.collectArrayListAligned()`
 
 ### Changed
 
@@ -30,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove incorrect overrides for `RepeatWith`
 - Remove incorrect overrides for `Cycle`
 - Remove `min()` and `max()` from `math_extra.zig` in favor of `@min()` and `@max()`
+- Remove `Iter.collect()`
 
 ### Fixed